fix: Resolve MSVC signed/unsigned and narrowing warnings across codebase
- Cast -1 sentinel values to static_cast<uint64_t>(-1) in ELF/PE/MachO/Strategy handlers
- Fix int-to-uint16_t narrowing in opcode_table OID computation
- Fix size_t-to-uint16_t truncation in instruction_t hash return
- Fix uint16_t-to-uint8_t narrowing in file_type_parser PE mode lookup
- Use explicit uint16_t keys in ELF/PE arch lookup tables
- Rename shadowed variable in ELFHandler::getEntryIndex

+++ sdk/src/segmentator/Strategy.cpp
@@ -12,7 +12,7 @@ using namespace VMPilot::SDK::Segmentator;
 std::pair<uint64_t, uint64_t>
 FileHandlerStrategy::doGetBeginEndAddr() noexcept {
     spdlog::error("FileHandlerStrategy::doGetBeginEndAddr not implemented");
-    return std::make_pair(-1, -1);
+    return std::make_pair(static_cast<uint64_t>(-1), static_cast<uint64_t>(-1));
 }
 
 std::vector<uint8_t> FileHandlerStrategy::doGetTextSection() noexcept {
@@ -22,7 +22,7 @@ std::vector<uint8_t> FileHandlerStrategy::doGetTextSection() noexcept {
 
 uint64_t FileHandlerStrategy::doGetTextBaseAddr() noexcept {
     spdlog::error("FileHandlerStrategy::doGetTextBaseAddr not implemented");
-    return -1;
+    return static_cast<uint64_t>(-1);
 }
 
 NativeSymbolTable FileHandlerStrategy::doGetSymbols() noexcept {
